# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Influencer Travel Marketplace — AI-Powered Itinerary Planning API
  description: |
    ## Overview

    REST API for the **Influencer Travel Marketplace** platform.
    This API enables three user roles to interact with the marketplace:

    | Role | Capabilities |
    |------|-------------|
    | `influencer` | Create and manage travel packages, upload media, use AI itinerary suggestions |
    | `admin` | Review, approve, and reject packages submitted for marketplace listing |
    | `customer` | Browse the marketplace, search for packages, view package details |

    ## Architecture

    - **Backend:** FastAPI (Python), deployed on Railway *— Phase 2 target; Phase 1 is Flask 3.0.3 on port 5000 with no `/v1` prefix. See [apps/api/docs/adr/0001-flask-now-fastapi-for-ai.md](../apps/api/docs/adr/0001-flask-now-fastapi-for-ai.md).*
    - **Database:** Supabase PostgreSQL (13 tables)
    - **AI:** LangChain RAG pipeline with GPT-4o (NFR-P-101: responses < 10 seconds)
    - **Frontend:** Next.js + Tailwind CSS, deployed on Vercel
    - **Authentication:** JWT Bearer tokens issued on login

    ## Package Lifecycle

    ```
    draft → pending_review → approved → live → archived
                          ↘ rejected → draft (via revise)
    ```

    ## Non-Functional Requirements

    - **NFR-P-100:** All user-facing text is localised to English (Australia); prices in whole AUD dollars.
...

    ## Assessor Notes

...

  version: "2.0.0"
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /auth/login`. Include in every protected request:
        ```
        Authorization: Bearer <token>
        ```
        The token payload contains:
        ```json
        {
          "sub": "uuid",
          "role": "influencer | admin | customer",
          "email": "user@example.com",
          "exp": 1748822400
        }
        ```

  # ─────────────────────────────────────────────
  #  REUSABLE SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    TravelPackageSummary:
      type: object
      properties:
        package_id:
          type: string
          format: uuid
          example: "pkg-00001-uuid-example-here"
        title:
          type: string
          example: "10-Day Bali Escape: Culture, Surf & Cuisine"
        destination_country:
          type: string
          nullable: true
          example: "Indonesia"
        destination_city:
          type: string
          nullable: true
          example: "Bali"
        duration_days:
          type: integer
          nullable: true
          example: 10
        base_price_aud:
          type: integer
          example: 3500
        status:
          $ref: '#/components/schemas/PackageStatus'
        creator_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        submitted_at:
          type: string
          format: date-time
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        cover_image_url:
          type: string
          format: uri
          nullable: true
          example: "https://storage.supabase.co/object/public/media/pkg-00001/cover.jpg"
    TravelPackageDetail:
      allOf:
        - $ref: '#/components/schemas/TravelPackageSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            max_group_size:
              type: integer
              nullable: true
            tags:
              type: array
              items:
                type: string
            flights:
              type: array
              items:
                $ref: '#/components/schemas/FlightDetail'
            hotels:
              type: array
              items:
                $ref: '#/components/schemas/HotelDetail'
            activities:
              type: array
              items:
                $ref: '#/components/schemas/ActivityDetail'
            media:
              type: array
              items:
                $ref: '#/components/schemas/MediaItem'
            days:
              type: array
              description: Per-day narrative for detail-page rendering, ordered by day_number. Group flights/hotels/activities client-side via their day_number / check_in_day fields.
              items:
                $ref: '#/components/schemas/PackageDay'
            creator:
              $ref: '#/components/schemas/PackageCreator'
              nullable: true
            latest_approval:
              $ref: '#/components/schemas/ApprovalRecord'
              nullable: true
            pricing:
              $ref: '#/components/schemas/Pricing'
              nullable: true
...
    PackageStatus:
      type: string
      enum: [draft, pending_review, approved, rejected, live, archived]
      description: |
        Package lifecycle states:
        - `draft` — being edited by influencer
        - `pending_review` — submitted, awaiting admin decision
        - `approved` — admin approved, ready to publish
        - `rejected` — admin rejected; influencer can revise and resubmit
        - `live` — published and visible in marketplace
        - `archived` — removed from marketplace
...
    FlightDetail:
      allOf:
        - type: object
          properties:
            flight_id:
              type: string
              format: uuid
            sequence_order:
              type: integer
              example: 1
        - $ref: '#/components/schemas/FlightInput'
...
    HotelDetail:
      allOf:
        - type: object
          properties:
            hotel_id:
              type: string
              format: uuid
            sequence_order:
              type: integer
              example: 1
        - $ref: '#/components/schemas/HotelInput'
...
    ActivityDetail:
      allOf:
        - type: object
          properties:
            activity_id:
              type: string
              format: uuid
            sequence_order:
              type: integer
              example: 1
        - $ref: '#/components/schemas/ActivityInput'

    # ── Approvals ─────────────────────────────
    ApprovalRecord:
      type: object
      properties:
        approval_id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, rejected]
          example: "rejected"
        rejection_reason:
          type: string
          nullable: true
          example: "Price does not reflect the stated inclusions. Please itemise all costs clearly."
        reviewed_at:
          type: string
          format: date-time
          example: "2025-03-10T14:22:00+10:00"
    ApproveRequest:
      type: object
      description: Body for approving a package. No required fields.
      properties:
        notes:
          type: string
          nullable: true
          description: Optional internal notes visible only to admins.
          example: "Verified flight prices against current market rates."
    RejectRequest:
      type: object
      required: [rejection_reason]
      properties:
        rejection_reason:
          type: string
          minLength: 10
          description: |
            Required explanation shown to the influencer so they can revise the package.
            Minimum 10 characters.
          example: "Price does not reflect the stated inclusions. Please itemise all costs clearly."
        notes:
          type: string
          nullable: true
          description: Optional internal admin notes (not shown to influencer).

    # ── Media ─────────────────────────────────
    MediaItem:
      type: object
      properties:
        media_id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        media_type:
          type: string
          enum: [image, video]
          example: "image"
        url:
          type: string
          format: uri
          example: "https://storage.supabase.co/object/public/media/pkg-00001/bali-rice-terrace.jpg"
        thumbnail_url:
          type: string
          format: uri
          nullable: true
          example: "https://storage.supabase.co/object/public/media/pkg-00001/bali-rice-terrace-thumb.jpg"
        caption:
          type: string
          nullable: true
          example: "Rice terraces at sunrise"
        is_cover:
          type: boolean
          description: Whether this item is the cover/thumbnail for the package.
          example: true
        sort_order:
          type: integer
          example: 1
        uploaded_at:
          type: string
          format: date-time

    # ── Detail-page extras ─────────────────────
...
    AISuggestRequest:
      type: object
      required: [package_id, prompt]
      properties:
        package_id:
          type: string
          format: uuid
          description: The draft package to generate suggestions for.
          example: "pkg-00001-uuid-example-here"
        prompt:
          type: string
          minLength: 10
          maxLength: 1000
          description: |
            Free-text prompt describing what the influencer wants suggestions for.
            The RAG pipeline uses existing package data as context.
          example: "Suggest a 3-day extension to explore the Gili Islands from Bali, including snorkelling and sunset dinner activities."
    AISuggestion:
      type: object
      properties:
        suggestion_id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        prompt:
          type: string
          example: "Suggest a 3-day extension to explore the Gili Islands..."
        suggestion_text:
          type: string
          description: Full structured itinerary suggestion from GPT-4o.
          example: |
            **Day 11 – Transfer to Gili Trawangan**
            - Morning: Fast boat from Padang Bai (2 hrs, ~AUD $35)
            - Afternoon: Check in at Ombak Sunset Resort; afternoon snorkel at Turtle Point
            - Evening: Sunset dinner at Scallywags Beach Club

            **Day 12 – Gili Air**
            - Morning: Bike ride around Gili Trawangan (circumference ~7 km)
            - Afternoon: Island hop to Gili Air by local cidomo horse cart boat
            - Sunset: Cocktails at Mowies Bar with Lombok volcano views

            **Day 13 – Return to Bali**
            - Morning: Stand-up paddleboarding or freediving lesson (~AUD $60)
            - Afternoon: Fast boat back to Bali; transfer to Ngurah Rai Airport
        status:
          type: string
          enum: [pending, accepted, dismissed]
          description: |
            - `pending` — generated, awaiting influencer decision
            - `accepted` — influencer accepted; components added to package
            - `dismissed` — influencer dismissed the suggestion
          example: "pending"
        generated_at:
          type: string
          format: date-time
        accepted_at:
          type: string
          format: date-time
          nullable: true
        response_time_ms:
          type: integer
          description: |
            Actual AI pipeline response time in milliseconds.
            NFR-P-101 requires this to be < 10,000.
          example: 4823
    AISuggestionAcceptRequest:
      type: object
      description: Accepting a suggestion can optionally auto-add itinerary components.
      properties:
        auto_apply:
          type: boolean
          default: true
          description: |
            If `true`, the backend parses `suggestion_text` and creates corresponding
            flight/hotel/activity records on the package. Set to `false` to mark the
            suggestion accepted without modifying the package.
          example: true

    # ── Marketplace ───────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items matching the query.
          example: 142
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8

    # ── Errors ────────────────────────────────
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: Machine-readable error identifier.
          example: "PACKAGE_NOT_FOUND"
        message:
          type: string
          description: Human-readable error description (English, Australia).
          example: "The requested travel package could not be found."
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional context (e.g. validation field errors).
          example:
            field: "base_price_aud"
            issue: "Must be a positive integer (whole AUD dollars)."

  # ─────────────────────────────────────────────
  #  REUSABLE PARAMETERS
  # ─────────────────────────────────────────────
  parameters:
    PackageIdPath:
      name: package_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the travel package.
      example: "pkg-00001-uuid-example-here"
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for paginated results.
    PerPageQuery:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page.

  # ─────────────────────────────────────────────
  #  REUSABLE RESPONSES
  # ─────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "UNAUTHORIZED"
            message: "A valid Bearer token is required to access this resource."
    Forbidden:
      description: Authenticated user does not have the required role for this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "FORBIDDEN"
            message: "Your account role does not permit this action."
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "NOT_FOUND"
            message: "Resource not found."
    UnprocessableEntity:
      description: Request body failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "VALIDATION_ERROR"
            message: "One or more fields are invalid."
            details:
              field: "base_price_aud"
              issue: "Must be a positive integer (whole AUD dollars)."


# ─────────────────────────────────────────────
#  PATHS
# ─────────────────────────────────────────────
paths:
...
  /approvals:
    get:
      tags: [approvals]
      summary: List packages pending admin review
      description: |
        Returns a paginated list of packages in `pending_review` status, ordered by
        `submitted_at` ascending (oldest submissions first, to prevent starvation).

        **Roles:** `admin`
        **FR:** FR-A-100
      operationId: listPendingApprovals
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - name: sort
          in: query
          schema:
            type: string
            enum: [submitted_at_asc, submitted_at_desc]
            default: submitted_at_asc
      responses:
        "200":
          description: Packages awaiting review.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TravelPackageSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /approvals/{package_id}/approve:
    parameters:
      - $ref: '#/components/parameters/PackageIdPath'
    post:
      tags: [approvals]
      summary: Approve a package for marketplace listing
      description: |
        Approves a package in `pending_review` status, transitioning it to `approved`.

        After approval the influencer may publish the package (transition to `live`)
        via a separate action. The approval record is stored in `package_approvals`
        and linked to the reviewing admin.

        **Roles:** `admin`
        **FR:** FR-A-101
      operationId: approvePackage
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
            example:
              notes: "Verified flight prices against current market rates. Content is high quality."
      responses:
        "200":
          description: Package approved. Status is now `approved`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  package:
                    $ref: '#/components/schemas/TravelPackageSummary'
                  approval:
                    $ref: '#/components/schemas/ApprovalRecord'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Package is not in `pending_review` status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "INVALID_STATUS_TRANSITION"
                message: "Only packages in 'pending_review' status can be approved. Current status: 'approved'."
  /approvals/{package_id}/reject:
    parameters:
      - $ref: '#/components/parameters/PackageIdPath'
    post:
      tags: [approvals]
      summary: Reject a package and return it to the influencer
      description: |
        Rejects a package in `pending_review` status, transitioning it to `rejected`.

        The `rejection_reason` is stored in `package_approvals` and surfaced to the
        influencer when they view the package. The influencer may then revise and
        resubmit (FR-I-103).

        **Roles:** `admin`
        **FR:** FR-A-102
      operationId: rejectPackage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
            example:
              rejection_reason: "Price does not reflect the stated inclusions. Please itemise all costs clearly and ensure the total matches `base_price_aud`."
              notes: "Comparison: equivalent Qantas holidays package for same dates is AUD $2,800."
      responses:
        "200":
          description: Package rejected. Status is now `rejected`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  package:
                    $ref: '#/components/schemas/TravelPackageSummary'
                  approval:
                    $ref: '#/components/schemas/ApprovalRecord'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Package is not in `pending_review` status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "INVALID_STATUS_TRANSITION"
                message: "Only packages in 'pending_review' status can be rejected."
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /approvals/{package_id}/publish:
    parameters:
      - $ref: '#/components/parameters/PackageIdPath'
    post:
      tags: [approvals]
      summary: Publish an approved package to the live marketplace
      description: |
        Transitions an `approved` package to `live`, making it visible in the
        public marketplace. Sets `published_at` timestamp.

        This action is available to both the owning influencer (self-publish)
        and admins.

        **Roles:** `influencer` (own packages), `admin`
        **FR:** FR-I-102
      operationId: publishPackage
      responses:
        "200":
          description: Package is now live in the marketplace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPackageSummary'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Package is not in `approved` status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "INVALID_STATUS_TRANSITION"
                message: "Only packages in 'approved' status can be published. Current status: 'draft'."

  # ══════════════════════════════════════════
  #  MEDIA
  # ══════════════════════════════════════════
...
  /ai/suggest:
    post:
      tags: [ai]
      summary: Generate an AI itinerary suggestion for a package
      description: |
        Submits a natural-language prompt to the LangChain RAG pipeline (GPT-4o backend)
        to generate a structured itinerary suggestion. The existing package data
        (destination, duration, existing flights/hotels/activities) is used as
        retrieval context.

        The suggestion is stored in the `ai_suggestions` table with status `pending`.
        Use `PATCH /ai/suggestions/{suggestion_id}/accept` to apply it to the package.

        **Performance:** NFR-P-101 — This endpoint must respond within **10 seconds**.
        The `response_time_ms` field in the response records actual latency.

        **Roles:** `influencer` (own packages only)
        **FR:** FR-I-105
      operationId: generateAISuggestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AISuggestRequest'
            example:
              package_id: "pkg-00001-uuid-example-here"
              prompt: "Suggest a 3-day extension to explore the Gili Islands from Bali, including snorkelling and sunset dinner activities suitable for couples."
      responses:
        "201":
          description: Suggestion generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISuggestion'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "504":
          description: AI pipeline timed out (exceeded NFR-P-101 10-second SLA).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "AI_TIMEOUT"
                message: "The AI suggestion pipeline did not respond within the 10-second time limit. Please try again."
  /ai/suggestions/{package_id}:
    parameters:
      - $ref: '#/components/parameters/PackageIdPath'
    get:
      tags: [ai]
      summary: List AI suggestions for a package
      description: |
        Returns all AI-generated suggestions for a given package, ordered by
        `generated_at` descending (most recent first).

        **Roles:** `influencer` (own packages only)
        **FR:** FR-I-105
      operationId: listAISuggestions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, dismissed]
          description: Filter suggestions by status.
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
      responses:
        "200":
          description: List of AI suggestions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AISuggestion'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /ai/suggestions/{suggestion_id}/accept:
    parameters:
      - name: suggestion_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Unique identifier of the AI suggestion.
    patch:
      tags: [ai]
      summary: Accept an AI suggestion and optionally apply it to the package
      description: |
        Marks a suggestion as `accepted` and, when `auto_apply` is true (default),
        parses the structured suggestion text and creates corresponding
        flight/hotel/activity records on the package.

        The package must be in `draft` or `rejected` status for `auto_apply` to work.
        Setting `auto_apply: false` marks the suggestion accepted without modifying
        the package — useful when the influencer copies content manually.

        **Roles:** `influencer` (own packages only)
        **FR:** FR-I-106
      operationId: acceptAISuggestion
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AISuggestionAcceptRequest'
            example:
              auto_apply: true
      responses:
        "200":
          description: Suggestion accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestion:
                    $ref: '#/components/schemas/AISuggestion'
                  package:
                    $ref: '#/components/schemas/TravelPackageDetail'
                    description: Present when `auto_apply` is true; reflects newly added components.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Suggestion is not in `pending` status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "SUGGESTION_ALREADY_RESOLVED"
                message: "This suggestion has already been accepted or dismissed."
  /ai/suggestions/{suggestion_id}/dismiss:
    parameters:
      - name: suggestion_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [ai]
      summary: Dismiss an AI suggestion
      description: |
        Marks a suggestion as `dismissed`. The suggestion is retained in the database
        for audit purposes but will not appear in the default `pending` filter.

        **Roles:** `influencer` (own packages only)
        **FR:** FR-I-105
      operationId: dismissAISuggestion
      responses:
        "200":
          description: Suggestion dismissed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISuggestion'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════════════════
  #  MARKETPLACE
  # ══════════════════════════════════════════
...